+++ source/isaaclab/docs/CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog
 ---------
 
+0.46.2 (2025-09-11)
+~~~~~~~~~~~~~~~~~~~
+
+Added
+^^^^^
+
+* Modified setter to support for viscous and dynamic joint friction coefficients in articulation based on IsaacSim 5.0.
+* Added randomization of viscous and dynamic joint friction coefficients in event term.
+
+
 0.46.1 (2025-09-10)
 ~~~~~~~~~~~~~~~~~~~
 
@@ -20,7 +30,6 @@ Added
   :meth:`~isaaclab.sim.utils.get_first_matching_child_prim` to control whether to traverse instance prims
   during the traversal. Earlier, instanced prims were skipped since :meth:`Usd.Prim.GetChildren` did not return
   instanced prims, which is now fixed.
-
 
 Changed
 ^^^^^^^